Exclude event discarded update! from transaction

+++ app/jobs/process_incoming_f_ticks_events.rb
@@ -9,14 +9,11 @@ class ProcessIncomingFTicksEvents
     FederatedLoginEvent.transaction do
       incoming_f_ticks_events.each do |incoming|
         subject = FederatedLoginEvent.new
-
-        if subject.create_instance(incoming.data)
-          incoming.destroy!
-        else
-          incoming.update!(discarded: true)
-        end
+        incoming.destroy! if subject.create_instance(incoming.data)
       end
     end
+
+    incoming_f_ticks_events.each { |event| event.update! discarded: true }
   end
 
   def incoming_f_ticks_events
